refactoring gemini flash model obj
