+++ src/events/actions/CreateNewChat.ts
@@ -1,7 +1,8 @@
+import 'dotenv/config';
 import { ChannelType, Message, } from "discord.js";
 
+import { GeminiFlash } from "../../llm/Gemini/GeminiFlash";
 import { getEnvVar } from "../../utils";
-import { callGeminiFlash } from "../../llm/Gemini/callGeminiFlash";
 
 import type { OmitPartialGroupDMChannel } from "discord.js";
 
@@ -29,7 +30,7 @@ export default async function CreateNewChat(msg: OmitPartialGroupDMChannel<Messa
         });
 
         const originalMessage = await thread.send("Waiting for AI to respond...");
-        await callGeminiFlash({
+        await GeminiFlash.addToQueue({
             threadMessage: originalMessage,
             prompt: prompt,
             chatHistory: [],
